Remove RollOptions component
